Main group 9 customer checkout

+++ back-end/src/database/migrations/20220804194029-sales.js
@@ -27,7 +27,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(6, 2),
       },
       delivery_address: {
         allowNull: false,
